refactor: reorganize API service structure and update import paths for better clarity
