+++ bmusic-ui/src/app/home/albums/albums.component.ts
@@ -4,7 +4,7 @@ import { AlbumStateService } from '../../service/states/album.state.service';
 import { Albums } from '../../service/models/album.interface';
 import { environment } from '../../../environments/environment';
 import { Router, RouterLink } from '@angular/router';
-import { ApiService } from '../../service/api.service';
+import { ApiService } from '../../service/apiCalls/api.service';
 
 @Component({
   selector: 'app-albums',
