+++ bmusic-ui/src/app/navbar/upload/upload.component.ts
@@ -1,10 +1,7 @@
-import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { ChangeDetectorRef, Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ApiService } from '../../service/api.service';
 import { HttpEventType } from '@angular/common/http';
-import { AlbumStateService } from '../../service/states/album.state.service';
-import { SongsStateService } from '../../service/states/songs.state.service';
-import { ArtistStateService } from '../../service/states/artist.state.service';
+import { ApiUploadService } from '../../service/apiCalls/api-upload.service';
 
 @Component({
   selector: 'app-upload',
@@ -13,23 +10,16 @@ import { ArtistStateService } from '../../service/states/artist.state.service';
   templateUrl: './upload.component.html',
   styleUrls: ['./upload.component.scss'],
 })
-export class UploadComponent implements OnInit {
+export class UploadComponent {
   selectedFiles: File[] = [];
   fileCount: number = 0;
   uploadProgress: number = 0;
   successMessage: string = '';
 
   constructor(
-    private api: ApiService,
-    private albumState: AlbumStateService,
-    private songsState: SongsStateService,
-    private artistState: ArtistStateService,
-    private cd: ChangeDetectorRef
+    private cd: ChangeDetectorRef,
+    private uploadService: ApiUploadService
   ) {}
-
-  ngOnInit() {
-    // No need to manually attach listeners; Angular handles (change) events.
-  }
 
   handleFileInputChange(event: Event) {
     const target = event.target as HTMLInputElement;
@@ -43,7 +33,7 @@ export class UploadComponent implements OnInit {
 
   uploadFiles() {
     if (this.selectedFiles.length > 0) {
-      this.api.uploadFiles(this.selectedFiles).subscribe({
+      this.uploadService.uploadFiles(this.selectedFiles).subscribe({
         next: (event) => {
           if (event.type === HttpEventType.UploadProgress && event.total) {
             this.uploadProgress = Math.round(
@@ -56,17 +46,6 @@ export class UploadComponent implements OnInit {
             this.successMessage = 'Files uploaded successfully!';
             this.uploadProgress = 0;
             this.cd.markForCheck();
-
-            // Fetch updated albums, songs, and artists
-            this.api.fetchAlbums().subscribe((albums) => {
-              this.albumState.setAlbums(albums);
-            });
-            this.api.fetchSongs().subscribe((songs) => {
-              this.songsState.setSongs(songs);
-            });
-            this.api.fetchArtists().subscribe((artists) => {
-              this.artistState.setArtists(artists);
-            });
           }
         },
         error: (error) => {
